you can now dismiss the create view with an x-button
