+++ Sources/WishKit/SwiftUI/WishlistView.swift
@@ -95,7 +95,10 @@ struct WishlistView: View {
                     AddButton(buttonAction: createWishAction)
                         .padding([.bottom, .trailing], 20)
                         .sheet(isPresented: $showingCreateSheet) {
-                            CreateWishView(createActionCompletion: { wishModel.fetchList() })
+                            CreateWishView(
+                                createActionCompletion: { wishModel.fetchList() },
+                                closeAction: { self.showingCreateSheet = false }
+                            )
                                 .frame(minWidth: 500, idealWidth: 500, minHeight: 400, maxHeight: 600)
                                 .background(backgroundColor)
                         }
